+++ doc/whats-new.rst
@@ -31,6 +31,10 @@ New Features
     ``pip install git+https://github.com/andrewgsavage/pint.git@refs/pull/6/head)``.
     Even with it, interaction with non-numpy array libraries, e.g. dask or sparse, is broken.
 
+- :py:func:`~xarray.dot`, and :py:func:`~xarray.DataArray.dot` now support the
+  ``dims=xarray.ALL_DIMS`` option to sum over the union of dimensions of all 
+  arrays (:issue:`3423`) by `Mathias Hauser <https://github.com/mathause>`_.
+
 Bug fixes
 ~~~~~~~~~
 - Fix regression introduced in v0.14.0 that would cause a crash if dask is installed